Modal checks helper added and basic tests for multiple modal usage started
