+++ tests/integration/options-test.js
@@ -2,6 +2,13 @@ import defaultFor from 'ember-modals/utils/default-for';
 import Em from 'ember';
 import { test } from 'ember-qunit';
 import startApp from '../helpers/start-app';
+import {
+  checkController,
+  checkModel,
+  checkTemplate,
+  checkView,
+  testModal
+} from '../helpers/checks';
 
 /* QUnit helpers */
 
@@ -142,100 +149,3 @@ variationsToTest.forEach(function(variation) {
     });
   });
 });
-
-
-/* Helper methods */
-
-var testModal = function(options, renderingOptions) {
-
-  /* Allow us to pass no object */
-
-  options = defaultFor(options, {});
-  renderingOptions = defaultFor(renderingOptions, {});
-
-  /* Then check each component of the modal */
-
-  checkTemplate(options.template);
-  checkView(options.view);
-  checkModel(options.model);
-  checkController(options.controller);
-
-  checkOutlet(renderingOptions.outlet);
-  checkParentView(renderingOptions.parentView);
-};
-
-/* Each of the below methods checks the given component of the
-modal. If an option is passed, it checks that the given option
-is set on the modal. If no option is passed it checks that the
-default option is set on the modal */
-
-var checkTemplate = function(name) {
-  var expectedName, text;
-
-  name = defaultFor(name, template);
-  expectedName = name.replace('/', '.');
-  text = name ? name : 'default';
-
-  equal(inspect('template_name').text().trim(), expectedName,
-    'Modal should have the ' + text + ' template');
-
-};
-
-var checkView = function(name) {
-  var expectedName;
-
-  name = defaultFor(name, defaultView);
-  expectedName = ':' + name + ':';
-
-  ok(inspect('view_constructor').text().indexOf(expectedName) > -1,
-    'Modal should have the ' + name + ' view');
-
-};
-
-var checkModel = function(expectedModel) {
-  var name;
-
-  if (expectedModel) {
-    name = expectedModel.name;
-
-    equal(inspect('model-name').text().trim(), name,
-      'Modal should have a model with name equal to ' + name);
-
-  } else {
-
-    ok(!inspect('model').is(':empty'),
-      'The modal should not have a model');
-
-  }
-
-};
-
-var checkController = function(name) {
-  var constructor = inspect('controller_constructor').text().trim();
-  var text;
-
-  name = defaultFor(name, currentRouteName());
-  text = name ? name : 'default (a.k.a. the route\'s controller';
-
-  ok(constructor.indexOf(name) > -1,
-    'Modal should have the ' + text + ' controller');
-
-};
-
-var checkOutlet = function(name) {
-  name = defaultFor(name, defaultOutlet);
-
-  equal(inspect('outlet_name').text().trim(), name,
-    'Modal should be rendered into the ' + name + ' outlet');
-
-};
-
-var checkParentView = function(name) {
-  var actualName = inspect('parent_view_name').text().trim();
-
-  name = defaultFor(name, defaultParentView);
-
-  equal(name, actualName,
-    'The modal\'s outlet should have the ' + name + ' parent view');
-
-};
